timer: switch from bash "time" to the original "time"
Previously I thought when I call "time" it will use the original
timing command (the one I can read about with "man time", but
actually it was a built in command in bash. It's missing all
the formatting capabilities of the original code, andy is
a bit annoying. To get rid of that, I have to call the original
time at is absolute path, that is usually /usr/bin/time.

Changes are made so the timing command can be easily switched to
other commands, and right now calling the original time.